requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==24.2.0
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
coverage==7.4.1
distlib==0.3.8
Django==4.2.9
django-cors-headers==4.3.1
django-environ==0.10.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
filelock==3.13.1
flake8==7.0.0
future==0.18.3
gunicorn==21.2.0
identify==2.5.35
idna==3.6
inflection==0.5.1
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
openapi-codec==1.3.2
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pre-commit==3.6.2
psycopg2-binary==2.9.9
punq==0.7.0
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
serializers==0.2.4
simplejson==3.19.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.25.0
